refactor: 중복 코드 제거를 위해 추상 클래스 활용함에 따른 DrawPolicy 제거
중복 코드를 위해 상속을 사용하는것은 지양해야하지만 완전한 is-A 관계이기 때문에 도입해도 상관없을 것이라 판단하였습니다
