+++ src/test/java/DealerTest.java
@@ -4,7 +4,6 @@ import domain.BlackjackDeckGenerator;
 import domain.CardValue;
 import domain.Dealer;
 import domain.Deck;
-import domain.DrawPolicy;
 import domain.Suit;
 import domain.TrumpCard;
 import domain.strategy.TestDrawStrategy;
@@ -23,7 +22,7 @@ public class DealerTest {
         Deque<TrumpCard> trumpCards = new LinkedList<>(
                 List.of(new TrumpCard(Suit.DIAMOND, CardValue.EIGHT), new TrumpCard(Suit.CLOVER, CardValue.NINE)));
         Deck deck = new Deck(new BlackjackDeckGenerator(), new TestDrawStrategy(trumpCards));
-        Dealer dealer = new Dealer(new DrawPolicy());
+        Dealer dealer = new Dealer();
         dealer.addDraw(deck.drawCard());
         dealer.addDraw(deck.drawCard());
         assertThat(dealer.isDrawable()).isFalse();
@@ -35,7 +34,7 @@ public class DealerTest {
                 List.of(new TrumpCard(Suit.DIAMOND, CardValue.EIGHT), new TrumpCard(Suit.DIAMOND, CardValue.J),
                         new TrumpCard(Suit.DIAMOND, CardValue.K)));
         Deck deck = new Deck(new BlackjackDeckGenerator(), new TestDrawStrategy(trumpCards));
-        Dealer dealer = new Dealer(new DrawPolicy());
+        Dealer dealer = new Dealer();
         assertThat(dealer.isDrawable()).isTrue();
     }
 }
